refactor: separate WebView2Host path rules from the control host
WebView2Host mixed two concerns: hosting a WPF WebView2 control, and deciding
where a host's Chromium user-data folder lives and which documents it may
navigate to. The second is pure path and URI logic, and WebView2HostTest only
ever exercised that, yet needed the WebView2 runtime to compile.

Moves BuildUserDataFolder, the orphan-folder sweep and IsAllowedDocumentUri into
WebView2Host.Paths.cs, which needs neither WebView2 nor WPF. Behaviour and
public API are unchanged.

SnykOptionsTest also joins the cross-platform run: it constructed a real
SnykSolutionService purely to satisfy a mock setup that nothing under test reads,
so it now uses a Mock<ISolutionService>, and its MockedVS collection attribute
moves to SnykOptionsTest.Vs.cs. Cross-platform run is now 412 tests.
