+++ Snyk.VisualStudio.Extension.Core.Tests/Snyk.VisualStudio.Extension.Core.props
@@ -62,6 +62,7 @@
     <Compile Include="$(SnykExtensionSourceRoot)UI\Html\TreeHtmlProvider.cs" Link="Core\UI\Html\TreeHtmlProvider.cs" />
     <Compile Include="$(SnykExtensionSourceRoot)UI\Html\WebView2BridgeBindings.cs" Link="Core\UI\Html\WebView2BridgeBindings.cs" />
     <Compile Include="$(SnykExtensionSourceRoot)UI\Html\WebView2EnvironmentCache.cs" Link="Core\UI\Html\WebView2EnvironmentCache.cs" />
+    <Compile Include="$(SnykExtensionSourceRoot)UI\Html\WebView2Host.Paths.cs" Link="Core\UI\Html\WebView2Host.Paths.cs" />
     <Compile Include="$(SnykExtensionSourceRoot)UI\Html\WebView2MessageDispatcher.cs" Link="Core\UI\Html\WebView2MessageDispatcher.cs" />
     <Compile Include="$(SnykExtensionSourceRoot)UI\Html\WebView2NavigationPreparer.cs" Link="Core\UI\Html\WebView2NavigationPreparer.cs" />
     <Compile Include="$(SnykExtensionSourceRoot)UI\Toolwindow\IHtmlPanel.cs" Link="Core\UI\Toolwindow\IHtmlPanel.cs" />
